Refactor reg_ex module location and update imports in bambooai.py
